requirements.txt
numpy==2.4.2
pandas==3.0.1
python-dotenv==1.2.2
Requests==2.32.5
scikit_learn==1.8.0
torch==2.10.0+cu130
xgboost==3.2.0